Renamed `index/zorder` package into `index/columnstats`
